refactor: fetch things from app
